did the styling for the home page and moved the login logic in another client component from the layout.tsx
