+++ frontend/src/components/item-card1.tsx
@@ -3,8 +3,8 @@ import { HeartIcon } from "@/assets/icons";
 
 export default function ItemCard() {
   return (
-    <div className="relative flex flex-col items-center gap-1">
-      <div className="w-36 h-40 relative">
+    <div className="relative col-span-6 sm:col-span-4 flex flex-col items-center gap-1">
+      <div className="w-full h-44 relative">
         <Image
           src="/blank-profile.png"
           alt="Cart Item"
@@ -15,7 +15,7 @@ export default function ItemCard() {
         />
       </div>
       <div className="flex justify-between w-full">
-        <p className="text-sm">Leather Jacket</p>
+        <p className="text-sm truncate">Leather Jacket</p>
         <p className="text-sm">$30</p>
       </div>
       <button className="absolute top-0 right-0 p-2 z-10">
